requirements.txt
docopt
colorama
cherrypy
requests
simplejson
sqlalchemy
argon2_cffi
psycopg2
